post_update

+++ src/brepnet/viz/find_nearest_pc_cd.py
@@ -40,7 +40,7 @@ if __name__ == "__main__":
     parser.add_argument("--prefix", type=str, required=False, default="")
     parser.add_argument("--txt", type=str, required=False, default="")
     parser.add_argument("--use_ray", action='store_true')
-    parser.add_argument("--num_gpus", type=float, default=1)
+    parser.add_argument("--num_gpus", type=int, default=1)
     parser.add_argument("--num_gpus_task", type=float, default=0.5)
 
     args = parser.parse_args()
